Develop

+++ app/src/main/java/com/mrostami/geckoin/model/RankedCoin.kt
@@ -59,5 +59,5 @@ data class RankedCoin(
     @SerializedName("total_supply")
     var totalSupply: Double? = null, // null
     @SerializedName("total_volume")
-    var totalVolume: Long? = null, // 14969737842
+    var totalVolume: Double? = null, // 14969737842
 )
